fix: get_images_from_sprite_sheet
In some cases, the function could return a sprite with a black background, regardless of the alpha value of the image. I'm not sure why, but maybe the clipping wasn't ideal. So I found this method online to use get_clip, set_clip and subsurface to extract spritesheets, and it keeps the original alpha intact.